requirements.txt
requests
django
uwsgi
mysqlclient
pillow
djangorestframework
django-cors-headers
djangorestframework-simplejwt